Move default createStorage into storage module
