+++ lib/storage.js
@@ -1,13 +1,17 @@
 var uint64be = require('uint64be')
 var flat = require('flat-tree')
+var raf = require('random-access-file')
 var createCache = require('./cache')
 
 module.exports = Storage
 
 var noarr = []
 
-function Storage (create, opts) {
-  if (!(this instanceof Storage)) return new Storage(create, opts)
+function Storage (createStorage, opts) {
+  if (!(this instanceof Storage)) return new Storage(createStorage, opts)
+
+  if (typeof createStorage === 'string') createStorage = defaultStorage(createStorage)
+  if (typeof createStorage !== 'function') throw new Error('Storage should be a function or string')
 
   const cache = createCache(opts)
 
@@ -19,7 +23,7 @@ function Storage (create, opts) {
   this._data = null
   this._bitfield = null
   this._signatures = null
-  this._create = create
+  this._create = createStorage
 }
 
 Storage.prototype.putData = function (index, data, nodes, cb) {
@@ -483,5 +487,14 @@ function readAll (st, offset, pageSize, cb) {
     if (err) return cb(null, bufs)
     bufs.push(buf)
     st.read(offset + bufs.length * pageSize, pageSize, loop)
+  }
+}
+
+function defaultStorage (dir) {
+  return function (name) {
+    try {
+      var lock = name === 'bitfield' ? require('fd-lock') : null
+    } catch (err) {}
+    return raf(name, { directory: dir, lock: lock })
   }
 }
